allow extra files for settings
There is $ZDOTDIR/zshenv.local now, which can take all kinds of settings
one wants that the default config does not give. Example of it is
zshenv.local.sample, copy over whatever you want.

Note that the zshenv.local file can not depend on any settings done later
in the config! For that see below.
That also means, if you for example, setup prompt colors there, ensure to
place the variables into '' so they are not expanded on config read, but
later!

For settings that depend on the zsh config to be loaded, use the file
$ZDOTDIR/.zshlate (only loaded in interactive shells) or the zsh standard
file $ZDOTDIR/.zlogin (only loaded in login shells)