zbell built into the prompt
Based on the zbell idea from Jean-Philippe Ouellet, see functions/zbell
for that, this is built into the prompt now. There are two zstyles to
control it, one to set the timeout (default 600), one to set the commands
to ignore (default '$EDITOR' '$PAGER' sleep man ssh, as well as the expanded
versions of the variables).

As all the variables and functions are renamed to fit in with all the rest of
the prompt, zbell and this can even be used together. zbells default time is
much shorter, so might make sense. Also, if the timeout is set to 0, none
of the prompt code is loaded, then you need zbell.