zbell prompt
add less to the default ignored commands.
also make drop_words really work, and drop any combination of
the to-be-dropped words from the command string.
That is, we remove the first occurence of
builtin, command, nocorrect, noglob, nohup and LANG=C,
in order to see the "real" command behind it.

Very helpful for aliases like alias man=nocorrect noglob man
which otherwise would beep - even when we dont want em too.