accept complete dirs
By default, filename completion examines all components of a path to
see if there are completions of that component, even if the component
matches an existing directory. ... When this style is ‘true’, any
prefix of a path that matches an existing directory is accepted
without any attempt to complete it further.